feat: refactor donut scheduler
